requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.4
djangorestframework==3.15.2
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.1
pytest-django==4.11.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2024.2
